Refactored calculated columns logic to XSKTableCreateProcessor
